# Optimize get_random_colors and convert_color