First, `get_random_colors`. List comprehensions are way faster
because you don't need to allocate the list multiple times.
For a million calls, it goes from 4.1s to 1.2s.

For `convert_color`, `sorted` actually allocates a list every time
whereas max/min is basic arithmetic. Using max/min is also way more
readable in what the code is actually doing (it took me a while to even
understand why it was sorting anything...). On its own, it makes
a million `get_type_colors` go from 4s to 3.44s.

With both those changes, it goes from 4.1s to 1.04s.

Worth noting that I checked that `convert_color` returns the exact same
thing for all values of rgb between -10 and 255 for each component.

```
get_random_colors | convert_color

old | old
4.116109441965818

old | new
3.4419355569407344

new | old
1.2401713710278273

new | new
1.043951882980764
```